+++ core/api/typescript/src/lib/enclaves/enclave_context.ts
@@ -19,7 +19,6 @@ import type {
     LoadModuleArgs,
     UnloadModuleArgs,
     GetModuleInfoArgs,
-    RegisterFilesArtifactsArgs,
     RegisterServiceArgs,
     GetServiceInfoArgs,
     WaitForHttpGetEndpointAvailabilityArgs,
@@ -36,7 +35,6 @@ import {
     newPartitionConnections,
     newPartitionServices,
     newPort,
-    newRegisterFilesArtifactsArgs,
     newRegisterServiceArgs,
     newRemoveServiceArgs,
     newRepartitionArgs,
@@ -50,7 +48,6 @@ import {
 } from "../constructor_calls";
 import type { ContainerConfig, FilesArtifactID } from "../services/container_config";
 import type { ServiceID } from "../services/service";
-import { SharedPath } from "../services/shared_path";
 import { ServiceContext } from "../services/service_context";
 import { PortProtocol, PortSpec } from "../services/port_spec";
 import type { GenericPathJoiner } from "./generic_path_joiner";
@@ -206,26 +203,9 @@ export class EnclaveContext {
     }
 
     // Docs available at https://docs.kurtosistech.com/kurtosis-core/lib-documentation
-    public async registerFilesArtifacts(filesArtifactUrls: Map<FilesArtifactID, string>): Promise<Result<null,Error>> {
-        const filesArtifactIdStrsToUrls: Map<string, string> = new Map();
-        for (const [artifactId, url] of filesArtifactUrls.entries()) {
-            filesArtifactIdStrsToUrls.set(String(artifactId), url);
-        }
-        const registerFilesArtifactsArgs: RegisterFilesArtifactsArgs = newRegisterFilesArtifactsArgs(filesArtifactIdStrsToUrls);
-
-        const registerFilesArtifactsResult = await this.backend.registerFilesArtifacts(registerFilesArtifactsArgs)
-
-        if(registerFilesArtifactsResult.isErr()){
-            return err(registerFilesArtifactsResult.error)
-        }
-        const result = registerFilesArtifactsResult.value
-        return ok(result)
-    }
-
-    // Docs available at https://docs.kurtosistech.com/kurtosis-core/lib-documentation
     public async addService(
             serviceId: ServiceID,
-            containerConfigSupplier: (ipAddr: string, sharedDirectory: SharedPath) => Result<ContainerConfig, Error>
+            containerConfigSupplier: (ipAddr: string) => Result<ContainerConfig, Error>
         ): Promise<Result<ServiceContext, Error>> {
 
         const resultAddServiceToPartition: Result<ServiceContext, Error> = await this.addServiceToPartition(
@@ -244,7 +224,7 @@ export class EnclaveContext {
     public async addServiceToPartition(
             serviceId: ServiceID,
             partitionId: PartitionID,
-            containerConfigSupplier: (ipAddr: string, sharedDirectory: SharedPath) => Result<ContainerConfig, Error>
+            containerConfigSupplier: (ipAddr: string) => Result<ContainerConfig, Error>
         ): Promise<Result<ServiceContext, Error>> {
 
         log.trace("Registering new service ID with Kurtosis API...");
@@ -263,10 +243,8 @@ export class EnclaveContext {
         const privateIpAddr: string = registerServiceResponse.getPrivateIpAddr();
         const relativeServiceDirpath: string = registerServiceResponse.getRelativeServiceDirpath();
 
-        const sharedDirectory = this.getSharedDirectory(relativeServiceDirpath)
-
         log.trace("Generating container config object using the container config supplier...")
-        const containerConfigSupplierResult: Result<ContainerConfig, Error> = containerConfigSupplier(privateIpAddr, sharedDirectory);
+        const containerConfigSupplierResult: Result<ContainerConfig, Error> = containerConfigSupplier(privateIpAddr);
         if (containerConfigSupplierResult.isErr()){
             return err(containerConfigSupplierResult.error);
         }
@@ -324,7 +302,6 @@ export class EnclaveContext {
         const serviceContext: ServiceContext = new ServiceContext(
             this.backend,
             serviceId,
-            sharedDirectory,
             privateIpAddr,
             privatePorts,
             startServiceResponse.getPublicIpAddr(),
@@ -357,14 +334,6 @@ export class EnclaveContext {
             );
         }
 
-        const relativeServiceDirpath: string = serviceInfo.getRelativeServiceDirpath();
-        if (relativeServiceDirpath === "") {
-            return err(new Error(
-                    "Kurtosis API reported an empty relative service directory path for service " + serviceId + " - this should never happen, and is a bug with Kurtosis!",
-                )
-            );
-        }
-
         const enclaveDataDirMountDirpathOnSvcContainer: string = serviceInfo.getEnclaveDataDirMountDirpath();
         if (enclaveDataDirMountDirpathOnSvcContainer === "") {
             return err(new Error(
@@ -372,8 +341,6 @@ export class EnclaveContext {
                 )
             );
         }
-
-        const sharedDirectory: SharedPath = this.getSharedDirectory(relativeServiceDirpath)
 
         const serviceCtxPrivatePorts: Map<string, PortSpec> = EnclaveContext.convertApiPortsToServiceContextPorts(
             serviceInfo.getPrivatePortsMap(),
@@ -385,7 +352,6 @@ export class EnclaveContext {
         const serviceContext: ServiceContext = new ServiceContext(
             this.backend,
             serviceId,
-            sharedDirectory,
             serviceInfo.getPrivateIpAddr(),
             serviceCtxPrivatePorts,
             serviceInfo.getPublicIpAddr(),
@@ -599,15 +565,6 @@ export class EnclaveContext {
     // ====================================================================================================
     //                                       Private helper functions
     // ====================================================================================================
-    private getSharedDirectory(relativeServiceDirpath: string): SharedPath {
-
-        const absFilepathOnThisContainer = this.pathJoiner.join(this.enclaveDataDirpath, relativeServiceDirpath);
-        const absFilepathOnServiceContainer = this.pathJoiner.join(SERVICE_ENCLAVE_DATA_DIR_MOUNTPOINT, relativeServiceDirpath);
-
-        const sharedDirectory = new SharedPath(absFilepathOnThisContainer, absFilepathOnServiceContainer, this.pathJoiner);
-        return sharedDirectory;
-    }
-
     private static convertApiPortsToServiceContextPorts(apiPorts: jspb.Map<string, Port>): Map<string, PortSpec> {
         const result: Map<string, PortSpec> = new Map();
         for (const [portId, apiPortSpec] of apiPorts.entries()) {